move type tests into its own folder
