+++ src/__typetest__/__fixtures__/dummy.d.ts
