+++ src/__typetest__/__fixtures__/pathString.ts
