+++ src/__typetest__/__fixtures__/traversable.ts
