+++ src/__typetest__/__fixtures__/tuple.ts
