+++ src/__typetest__/__fixtures__/type.ts
